+++ lib/slide_action.dart
@@ -1,0 +1,8 @@
+library slide_action;
+
+export 'src/slide_action_widget.dart'
+    show SlideAction, SlideActionWidgetBuilder;
+
+export 'src/slide_action_end_behavior.dart' show SlideActionEndBehavior;
+
+export 'src/slide_action_state_mixin.dart' show SlideActionStateMixin;