+++ lib/slide_to_perform.dart
@@ -1,8 +1,0 @@
-library slide_to_perform;
-
-export 'src/slide_to_perform.dart'
-    show
-        SlideToPerform,
-        SlideToPerformStateMixin,
-        SlideToPerformWidgetBuilder,
-        SlideToPerformEndBehavior;