+++ lib/src/frame_change_callback_provider.dart
@@ -1,18 +1,18 @@
+import 'package:flutter/animation.dart';
 import 'package:flutter/scheduler.dart';
 
-/// `delta` is the time elapsed since last frame.
+/// `delta` stores the duration elapsed since last frame.
 typedef FrameChangeCallback = void Function(Duration delta);
 
 /// Provides functionality for executing a function every frame
 /// in a given context.
 class FrameChangeCallbackProvider {
-
   /// Creates a [FrameChangeCallbackProvider].
-  /// 
+  ///
   /// * `vsync` is the ticker provider for current context.
   /// * `callback` the callback to execute whenever a frame changes. See [FrameChangeCallback].
-  /// 
-  /// ## Notes:
+  ///
+  /// ### Notes:
   ///
   /// * Use `start` and `stop` methods to control when to listen for changes.
   /// An instance of this class is in a *stopped* state on creation.
@@ -28,6 +28,7 @@ class FrameChangeCallbackProvider {
   final TickerProvider vsync;
 
   /// The callback to be fired every frame.
+  ///
   /// The [FrameChangeCallbackProvider] instance must be started for the callback to fire.
   final FrameChangeCallback callback;
 
@@ -41,7 +42,10 @@ class FrameChangeCallbackProvider {
   }
 
   /// Starts the [FrameChangeCallbackProvider] instance.
+  ///
   /// Does nothing if already running.
+  ///
+  /// Must not be called after the instance is disposed.
   void start() {
     if (_internalTicker.isActive) return;
 
@@ -50,14 +54,16 @@ class FrameChangeCallbackProvider {
   }
 
   /// Stops the [FrameChangeCallbackProvider] instance.
+  ///
   /// Does nothing if already stopped.
+  ///
+  /// Must not be called after the instance is disposed.
   void stop() {
     if (!_internalTicker.isActive) return;
     _internalTicker.stop();
   }
 
-  /// Call to dispose the [FrameChangeCallbackProvider] instance
-  /// if it is no longer in use.
+  /// Release the resources used by the instance. The instance is no longer usable after this method is called.
   void dispose() {
     _internalTicker.dispose();
   }